Move update schema to schema.go
Cut-paste, no code has changed
